pyproject.toml
[tool.poetry]
name = "hh"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
bs4 = "^0.0.1"
lxml = "^4.9.3"
fake-useragent = "^1.2.1"
openai = "^0.27.8"
python-telegram-bot = "^13.7"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
